no shared base class
