+++ ansible_navigator/cli.py
@@ -180,7 +180,7 @@ def setup_config() -> Tuple[List[str], NavigatorConfig]:
     cfg_env_var = "ANSIBLE_NAVIGATOR_CONFIG"
     env_config_path, msgs = env_var_is_file_path(cfg_env_var, "config")
     pre_logger_msgs += msgs
-    
+
     # Check well know locations
     found_config_path, msgs = get_conf_path(
         "ansible-navigator", allowed_extensions=["yml", "yaml", "json"]
@@ -217,7 +217,6 @@ def setup_config() -> Tuple[List[str], NavigatorConfig]:
                     error_and_exit_early(
                         "Config file at {0} but failed to parse it.".format(config_path)
                     )
-    # print(pre_logger_msgs)
 
     if config_path and config and config.get("ansible-navigator"):
         # If the config file was found and has the key we expect, log and use it
