+++ tests/integration/test_set_environment_variable.py
@@ -8,7 +8,6 @@ import pytest
 
 import ansible_navigator.cli as cli
 
-from .cli2runner import Cli2Runner
 from ..defaults import FIXTURES_DIR
 
 test_data = [
@@ -45,8 +44,13 @@ test_data = [
     argvalues=test_data,
     ids=[f"{idx}: {i[0]}" for idx, i in enumerate(test_data)],
 )
-class Test(Cli2Runner):
+class Test:
     # pylint: disable=too-few-public-methods
+    # pylint: disable=unused-argument
+    # pylint: disable=attribute-defined-outside-init
+    # pylint: disable=redefined-outer-name
+    # pylint: disable=too-many-arguments
+
     """test the use of set_environment_variable throguh to runner"""
 
     TEST_FIXTURE_DIR = f"{FIXTURES_DIR}/integration/set_environment_variable"
@@ -80,3 +84,53 @@ class Test(Cli2Runner):
 
         for item in expected.items():
             assert item in kwargs["envvars"].items()
+
+    @pytest.fixture(autouse=True)
+    def _setup(self):
+        self.cli_entry = "ansible-navigator {0} {1} -m {2}"
+
+    @mock.patch("ansible_navigator.runner.api.get_ansible_config")
+    def test_config_interactive(
+        self, mocked_runner, comment, cli_entry, config_fixture, expected, patch_curses
+    ):
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.INTERACTIVE["config"], cli_entry, "interactive")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
+
+    @mock.patch("ansible_navigator.runner.api.run_command")
+    def test_config_stdout(self, mocked_runner, comment, cli_entry, config_fixture, expected):
+        # pylint: disable=unused-argument
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.STDOUT["config"], cli_entry, "stdout")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
+
+    @mock.patch("ansible_navigator.runner.api.get_inventory")
+    def test_inventory_interactive(
+        self, mocked_runner, comment, cli_entry, config_fixture, expected, patch_curses
+    ):
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.INTERACTIVE["inventory"], cli_entry, "interactive")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
+
+    @mock.patch("ansible_navigator.runner.api.run_command")
+    def test_inventory_stdout(self, mocked_runner, comment, cli_entry, config_fixture, expected):
+        # pylint: disable=unused-argument
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.STDOUT["inventory"], cli_entry, "stdout")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
+
+    @mock.patch("ansible_navigator.runner.api.run_command_async")
+    def test_run_interactive(
+        self, mocked_runner, comment, cli_entry, config_fixture, expected, patch_curses
+    ):
+        # pylint: disable=unused-argument
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.INTERACTIVE["run"], cli_entry, "interactive")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
+
+    @mock.patch("ansible_navigator.runner.api.run_command_async")
+    def test_run_stdout(self, mocked_runner, comment, cli_entry, config_fixture, expected):
+        # pylint: disable=unused-argument
+        """test use of set_environment_variable"""
+        cli_entry = self.cli_entry.format(self.STDOUT["run"], cli_entry, "stdout")
+        self.run_test(mocked_runner, cli_entry, config_fixture, expected)
